Commit v0.642
-2 player mode working.
-Need to implement lifes/seperate scoring systems/keys

+++ Pump It Up Prime 2015/Assets/Scripts/PlayerPref.cs
@@ -15,7 +15,7 @@ public struct ScoreData {
 public struct PlayerSettings {
     public ScoreData playerScore;
     public float prefSpeed;
-    public int life;
+    public float life;
 }
 
 public static class PlayerPref {
